Improve build script

+++ CLI/build.ps1
@@ -1,2 +1,3 @@
+if (Test-Path .\artifact\) { Remove-Item -Recurse .\artifact\ }
 dotnet publish -c Publish -o artifact
-Remove-Item artifact\*.pdb
+if ($?) { Remove-Item artifact\*.pdb }
